+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated kit purchasing links. 
 - MIN_ERG_CADENCE created and changed from 20 to 30.
 - Fixed DNS server in AP mode.
+- Switched from using Power Table to a Torque Table for better compensation in cad variations. 
 
 ### Hardware
 - Wire diameter reduced from 7.2mm to 6.0mm on the window passthrough to accommodate the latest batch of cables.